`ServerRequestAttributeTrait` removed
